+++ tabela_simbolos.rb
@@ -1,0 +1,15 @@
+@tabela_simbolos = Array.new
+
+@tabela_simbolos.push(Token.new("inicio","inicio",nil))
+@tabela_simbolos.push(Token.new("varinicio","varinicio",nil))
+@tabela_simbolos.push(Token.new("varfim","varfim",nil))
+@tabela_simbolos.push(Token.new("escreva","escreva",nil))
+@tabela_simbolos.push(Token.new("leia","leia",nil))
+@tabela_simbolos.push(Token.new("se","se",nil))
+@tabela_simbolos.push(Token.new("entao","entao",nil))
+@tabela_simbolos.push(Token.new("fimse","fimse",nil))
+@tabela_simbolos.push(Token.new("faca-ate","faca-ate",nil))
+@tabela_simbolos.push(Token.new("fimfaca","inicio",nil))
+@tabela_simbolos.push(Token.new("inteiro","inteiro",nil))
+@tabela_simbolos.push(Token.new("lit","lit",nil))
+@tabela_simbolos.push(Token.new("real","real",nil))
