+++ token_class.rb
@@ -1,0 +1,12 @@
+class Token
+
+	def initialize(classe,lexema,tipo)
+		@classe = classe
+		@lexema = lexema
+		@tipo = tipo
+	end
+
+	attr_reader :classe
+	attr_reader :lexema
+	attr_reader :tipo
+end
